AGAT.LocoDispatcher.Business.Data/Helpers/LogicHelpers/EventHelper.cs
using AGAT.LocoDispatcher.Data.Managers;
using AGAT.LocoDispatcher.Data.Models.EventModels;
using System;
using System.Threading.Tasks;

namespace AGAT.LocoDispatcher.Data.Helpers.LogicHelpers
{
    public class EventHelper
    {
        private DataManager manager;
        public EventHelper()
        {
            manager = new DataManager();
        }

        public async Task<int> GetLocoShiftIdByLocoNumber(string locoNumber)
        {
            try
            {
                LocoShiftEvent shiftEvent = await manager.shiftRepository.GetActiveByNameAsync(locoNumber);
                if (shiftEvent is null)
                {
                    LocoShiftEvent newShiftEvent = new LocoShiftEvent
                    {
                        CreatedAt = DateTime.Now,
                        ESR = "",
                        IsValid = false,
                        StartShift = DateTime.Now,
                        TrainNumber = locoNumber
                    };
                    await manager.shiftRepository.CreateAsync(newShiftEvent);
                    return await GetLocoShiftIdByLocoNumber(locoNumber);
                }
                else
                {
                    return shiftEvent.Id;
                }
            }
            catch (Exception ex)
            {
                throw ex;
            }

        }
    }
}
